+++ src/pytestshellutils/downgraded/shell.py
@@ -283,10 +283,6 @@ class Factory(BaseFactory):
     """
     Base shell factory class.
 
-    :keyword pathlib.Path cwd:
-        The path to the desired working directory
-    :keyword dict environ:
-        A dictionary of ``key``, ``value`` pairs to add to the environment.
     :keyword bool slow_stop:
         Whether to terminate the processes by sending a :py:attr:`SIGTERM` signal or by calling
         :py:meth:`~subprocess.Popen.terminate` on the sub-process.
@@ -301,8 +297,6 @@ class Factory(BaseFactory):
         of ``_timeout`` instead of ``self.timeout``.
     """
 
-    cwd = attr.ib(converter=resolved_pathlib_path)
-    environ = attr.ib(repr=False)
     slow_stop = attr.ib(default=True)
     system_encoding = attr.ib(repr=False)
     timeout = attr.ib()